#74 - id 정보를 담고 있는 변수의 자료형을 Long으로 변환 + 더 이상 않는 메소드 삭제
- id 값은 int의 최대값보다 커질 수도 있기 때문에 Long 타입으로 변경해주었다.
- 더 이상 사용하지 않는 메소드들을 삭제했다.

+++ src/main/java/com/ncookie/imad/domain/contents/controller/ContentsController.java
@@ -39,7 +39,7 @@ public class ContentsController {
     @GetMapping("/details")
     public ApiResponse<TmdbDetails> getContentsDetails(
             @RequestHeader("Authorization") String accessToken,
-            @RequestParam(value = "id") int id,
+            @RequestParam(value = "id") Long id,
             @RequestParam(value = "type") String type) {
         ContentsType contentsType = ContentsType.valueOf(type.toUpperCase());
         // 이전에 저장했던 데이터라면 TMDB API 사용하지 않고 로컬 DB 조회하여 반환
