+++ src/main/java/com/ncookie/imad/domain/contents/service/ContentsService.java
@@ -35,11 +35,11 @@ public class ContentsService {
         return apiClient.searchByQuery(query, type, page);
     }
 
-    public TmdbDetails getContentsDetails(int id, String type) {
+    public TmdbDetails getContentsDetails(Long id, String type) {
         return apiClient.getContentsDetails(id, type);
     }
 
-    public String getContentsCertification(int id, String type) {
+    public String getContentsCertification(Long id, String type) {
         return apiClient.getContentsCertification(id, type);
     }
 
@@ -61,7 +61,7 @@ public class ContentsService {
      * =====================================
      */
 
-    public Contents getContentsByTmdbIdAndTmdbType(long id, ContentsType type) {
+    public Contents getContentsByTmdbIdAndTmdbType(Long id, ContentsType type) {
         return contentsRepository.findByTmdbIdAndTmdbType(id, type);
     }
 
@@ -79,7 +79,7 @@ public class ContentsService {
         return tvProgramDataRepository.save(tvProgramData);
     }
 
-    public boolean checkDuplicationExists(long id, ContentsType type) {
+    public boolean checkDuplicationExists(Long id, ContentsType type) {
         return contentsRepository.existsByTmdbIdAndTmdbType(id, type);
     }
